[SERVICE-543] Documentation, hiding some intents
