+++ src/client/main.ts
@@ -8,7 +8,7 @@ import {Identity} from 'openfin/_v2/main';
 
 import {tryServiceDispatch, getServicePromise} from './connection';
 import {Context} from './context';
-import {Application} from './directory';
+import {Application, AppName} from './directory';
 import {APIFromClientTopic, APIToClientTopic, RaiseIntentPayload, EventTransport} from './internal';
 import {ChannelChangedEvent, getChannelObject} from './contextChannels';
 import {parseContext, validateEnvironment} from './validation';
@@ -29,18 +29,31 @@ export * from './intents';
 export * from './errors';
 
 /**
- * Intent descriptor
+ * Contains the data you need to describe an intent.
  */
 export interface IntentMetadata {
+    /**
+     * The machine readable name of the intent.
+     */
     name: string;
+    /**
+     * The human-readable name of the intent.
+     */
     displayName: string;
 }
 
 /**
- * An interface that relates an intent to apps
+ * An interface that relates an intent to apps. This is returned by [[findIntent]] and [[findIntentsByContext]], which gives
+ * you a set of apps that can execute a particular intent.
  */
 export interface AppIntent {
+    /**
+     * Descriptor of this intent.
+     */
     intent: IntentMetadata;
+    /**
+     * An array of applications that are associated with this intent.
+     */
     apps: Application[];
 }
 
@@ -55,23 +68,32 @@ export interface AppIntent {
  * ```
  */
 export interface IntentResolution {
-    source: string;
+    /**
+     * The human-readable name of the app that resolved this intent.
+     */
+    source: AppName;
+    /**
+     * Any data returned by the intent.
+     */
     data?: unknown;
+    /**
+     * For future use. Right now always the string '1.0.0'.
+     */
     version: string;
 }
 
+/**
+ * Listener type alias. One of ContextListener or IntentListener.
+ */
 export type Listener = ContextListener | IntentListener;
 
+/**
+ * Listener for context broadcasts. Generated by [[addContextListener]].
+ */
 export interface ContextListener {
-    handler: (context: Context) => void;
     /**
-     * Unsubscribe the listener object.
+     * The callback for when this listener receives a context broadcast.
      */
-    unsubscribe: () => void;
-}
-
-export interface IntentListener {
-    intent: string;
     handler: (context: Context) => void;
     /**
      * Unsubscribe the listener object.
@@ -80,9 +102,32 @@ export interface IntentListener {
 }
 
 /**
- * Defines all events that are fired by the service
+ * Listener for intent sending. Generated by [[addIntentListener]].
+ */
+export interface IntentListener {
+    /**
+     * The intent name that we are listening to. Is whatever is passed into [[addIntentListener]].
+     */
+    intent: string;
+    /**
+     * The callback for when this listener receives an intent.
+     */
+    handler: (context: Context) => void;
+    /**
+     * Unsubscribe the listener object.
+     */
+    unsubscribe: () => void;
+}
+
+/**
+ * Defines all events that are fired by the service. Right now, The only event is the channel changed event.
+ * This will become a union type if more events are added.
  */
 export type FDC3Event = ChannelChangedEvent;
+/**
+ * Type to define the type of events fired by the service. Right now, the only event is the channel changed event.
+ * This will become a union type if more types are added.
+ */
 export type FDC3EventType = 'channel-changed';
 
 /**
@@ -119,6 +164,8 @@ const contextListeners: ContextListener[] = [];
  *     //with context
  *     agent.open('myApp', context);
  * ```
+ * @param name The app name to launch.
+ * @param context A context to pass to the app post-'launch.
  */
 export async function open(name: string, context?: Context): Promise<void> {
     return tryServiceDispatch(APIFromClientTopic.OPEN, {name, context: context && parseContext(context)});
@@ -146,6 +193,8 @@ export async function open(name: string, context?: Context): Promise<void> {
  * // raise the intent against a particular app
  * await agent.raiseIntent(appIntent.intent.name, context, appIntent.apps[0].name);
  * ```
+ * @param intent The intent name to find.
+ * @param context An optional context to send to find the intent.
  */
 export async function findIntent(intent: string, context?: Context): Promise<AppIntent> {
     return tryServiceDispatch(APIFromClientTopic.FIND_INTENT, {intent, context: context && parseContext(context)});
@@ -183,18 +232,20 @@ export async function findIntent(intent: string, context?: Context): Promise<App
  * // raise the intent, passing the given context, targeting the app
  * await agent.raiseIntent(startChat.intent.name, context, selectedApp.name);
  * ```
+ * @param context The context to send to find the intents that support it.
  */
 export async function findIntentsByContext(context: Context): Promise<AppIntent[]> {
     return tryServiceDispatch(APIFromClientTopic.FIND_INTENTS_BY_CONTEXT, {context: parseContext(context)});
 }
 
 /**
- * Publishes context to other apps on the desktop.
+ * Publishes context to other apps on the desktop. Any apps using [[addContextListener]] will receive this.
  * ```javascript
  *  agent.broadcast(context);
  * ```
  *
  * @throws `TypeError`: If `context` is not a valid Context
+ * @param context The context to broadcast.
  */
 export function broadcast(context: Context): void {
     tryServiceDispatch(APIFromClientTopic.BROADCAST, {context: parseContext(context)});
@@ -206,15 +257,22 @@ export function broadcast(context: Context): void {
  * //raise an intent to start a chat with a given contact
  * const intentR = await agent.findIntent("StartChat", context);
  * //use the IntentResolution object to target the same chat app with a new context
- * agent.raiseIntent("StartChat", newContext, intentR.source);
+ * agent.raiseIntent("StartChat", newContext, intentR.apps[0].name);
  * ```
+ * @param intent The intent name to raise.
+ * @param context The context that will be sent with this intent.
+ * @param target An optional [[AppName]] to send the intent to.
  */
-export async function raiseIntent(intent: string, context: Context, target?: string): Promise<IntentResolution> {
+export async function raiseIntent(intent: string, context: Context, target?: AppName): Promise<IntentResolution> {
     return tryServiceDispatch(APIFromClientTopic.RAISE_INTENT, {intent, context: parseContext(context), target});
 }
 
 /**
  * Adds a listener for incoming Intents from the Agent.
+ *
+ * To unsubscribe, use the returned [[IntentListener]].
+ * @param intent The name of the intent to listen to.
+ * @param handler The handler to call when we get sent an intent.
  */
 export function addIntentListener(intent: string, handler: (context: Context) => void): IntentListener {
     validateEnvironment();
@@ -248,6 +306,9 @@ export function addIntentListener(intent: string, handler: (context: Context) =>
 
 /**
  * Adds a listener for incoming context broadcast from the Desktop Agent.
+ *
+ * To unsubscribe, use the returned [[ContextListener]].
+ * @param handler The callback function to call when we receive a broadcast context.
  */
 export function addContextListener(handler: (context: Context) => void): ContextListener {
     validateEnvironment();
@@ -276,12 +337,23 @@ export function addContextListener(handler: (context: Context) => void): Context
  */
 export function addEventListener(eventType: 'channel-changed', handler: (event: ChannelChangedEvent) => void): void;
 
+/**
+ * Generic event listener. This is not for intent or context subscription: use [[addIntentListener]] and [[addContextListener]], respectively.
+ * @param eventType The event type.
+ * @param handler The callback when the event is fired.
+ * @param identity The OF window identity. Currently unused.
+ */
 export function addEventListener(eventType: FDC3EventType, handler: (event: FDC3Event) => void, identity?: Identity): void {
     validateEnvironment();
 
     eventEmitter.addListener(eventType, handler);
 }
 
+/**
+ * Remove any events you are subscribed to.
+ * @param eventType The type of the event to remove.
+ * @param handler The listener you had previously used to subscribe with.
+ */
 export function removeEventListener(eventType: FDC3EventType, handler: (eventPayload: ChannelChangedEvent) => void): void {
     validateEnvironment();
